+++ CHANGELOG.md
@@ -3,6 +3,25 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [1.3.2] - unrealeased
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+## [1.3.1] - 07/11/20
+### Added
+- ascii time output when users use ghorgignore
+- ascii time output when users use output-dir flag
+### Changed
+- default GHORG_ABSOLUTE_PATH_TO_CLONE_TO to $HOME/Desktop/ghorg
+### Deprecated
+### Removed
+### Fixed
+### Security
+
 ## [1.3.0] - 07/11/20
 ### Added
 - auto downcase of ghorg clone folder name; thanks @zamariola